wrote a function to process numeric strings
This function should break it up if the later part is a valid word. I also added some stuff to call the functions. If the stuff I just added but haven't tested works then I think this should function properly. might need to add a case for if the string is empty