Added UpdateList and an item event
I think that the updateList method should be ok now. The Item Event listener is really confusing, but I think we want it to have specific behavior if a file is selected. Still needs the logic for getting double clicked. The items in the list are just strings not connected to the files. So adding that behavior is going to be complicated